add json-web-token file and moved all jwt related stuff into it
